pyproject.toml
[tool.poetry]
name = "chronos"
version = "0.2.0"
description = "Voice RAG Orchestrator"
authors = ["Asif"]
package-mode = false

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
fastapi = "^0.111.0"
uvicorn = {extras = ["standard"], version = "^0.29.0"}
websockets = "^12.0"
python-dotenv = "^1.0.0"
pydantic-settings = "^2.2.0"

# --- Database & Authentication ---
sqlalchemy = {extras = ["asyncio"], version = "^2.0.0"}
asyncpg = "^0.29.0"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-jose = {extras = ["cryptography"], version = "^3.3.0"}

# --- AI & Logic ---
# We use LangGraph 0.2+ for the stable StateGraph API
langgraph = "^0.2.0"
langchain-core = "^0.3.0"
langchain-openai = "^0.3.0"
langchain-pinecone = "^0.2.0"
langchain-community = "^0.3.0"

# --- Document Processing ---
pypdf = "^3.17.0"
python-docx = "^1.1.0"
unstructured = "^0.11.0"
python-multipart = "^0.0.9"
docx2txt = "^0.8"

# --- Voice Services ---
# Deepgram v3.2+ is required for the Enum fix
deepgram-sdk = "^3.2.0"
openai = "^1.30.0"

# --- Data & Utilities ---
redis = "^5.0.4"
numpy = "^1.26.0"
watchfiles = "^0.21.0"

# --- Frontend (Streamlit) ---
streamlit = "^1.28.0"
pyaudio = "^0.2.14"
pydub = "^0.25.1"
requests = "^2.31.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"